# cortes-ia-contracts/design/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cortes IA — contrato proposto",
    "version": "0.2.0",
    "description": "Contrato-alvo atualizado pelas decisões v0.2. Não confundir com OpenAPI efetivamente gerado pela API inicial; ver API_IMPLEMENTADA.md."
  },
  "servers": [
    {
      "url": "https://api.example.invalid",
      "description": "Placeholder deliberado; substituir pelo host do ambiente."
    }
  ],
  "paths": {
...
    "/api/v1/purchases": {
      "post": {
        "operationId": "payments_create",
        "summary": "Criar compra Pix/cartão",
        "description": "Ownership e validação server-side; regras no plano técnico.",
        "tags": [
          "payments"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "201": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Purchase"
                }
              }
            }
          },
          "default": {
            "description": "Erro sanitizado; códigos e condições descritos em CONVENCOES.md.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 128
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseInput"
              }
            }
          }
        }
      },
      "get": {
        "operationId": "payments_list",
        "summary": "Histórico de compras",
        "description": "Ownership e validação server-side; regras no plano técnico.",
        "tags": [
          "payments"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchasePage"
                }
              }
            }
          },
          "default": {
            "description": "Erro sanitizado; códigos e condições descritos em CONVENCOES.md.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ]
      }
    },
...
    "/api/v1/tickets": {
      "post": {
        "operationId": "support_create",
        "summary": "Abrir chamado",
        "description": "Ownership e validação server-side; regras no plano técnico.",
        "tags": [
          "support"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "201": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            }
          },
          "default": {
            "description": "Erro sanitizado; códigos e condições descritos em CONVENCOES.md.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 128
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketInput"
              }
            }
          }
        }
      },
      "get": {
        "operationId": "support_list",
        "summary": "Listar chamados",
        "description": "Ownership e validação server-side; regras no plano técnico.",
        "tags": [
          "support"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketPage"
                }
              }
            }
          },
          "default": {
            "description": "Erro sanitizado; códigos e condições descritos em CONVENCOES.md.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ]
      }
    },
...
    "/api/v1/admin/users": {
      "get": {
        "operationId": "admin_users",
        "summary": "Pesquisar usuários",
        "description": "MFA e RBAC granular; ações sensíveis exigem step-up e motivo.",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountPage"
                }
              }
            }
          },
          "default": {
            "description": "Erro sanitizado; códigos e condições descritos em CONVENCOES.md.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          }
        ]
      }
    },
...
    "/api/v1/admin/users/{accountId}/block": {
      "post": {
        "operationId": "admin_block",
        "summary": "Bloquear/desbloquear conta",
        "description": "MFA e RBAC granular; ações sensíveis exigem step-up e motivo.",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ack"
                }
              }
            }
          },
          "default": {
            "description": "Erro sanitizado; códigos e condições descritos em CONVENCOES.md.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 128
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockInput"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/purchases": {
      "get": {
        "operationId": "admin_purchases",
        "summary": "Consultar compras",
        "description": "MFA e RBAC granular; ações sensíveis exigem step-up e motivo.",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchasePage"
                }
              }
            }
          },
          "default": {
            "description": "Erro sanitizado; códigos e condições descritos em CONVENCOES.md.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ]
      }
    },
    "/api/v1/admin/jobs": {
      "get": {
        "operationId": "admin_jobs",
        "summary": "Monitorar fila e erros",
        "description": "MFA e RBAC granular; ações sensíveis exigem step-up e motivo.",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminJobPage"
                }
              }
            }
          },
          "default": {
            "description": "Erro sanitizado; códigos e condições descritos em CONVENCOES.md.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          }
        ]
      }
    },
...
    "/api/v1/admin/tickets": {
      "get": {
        "operationId": "admin_tickets",
        "summary": "Consultar fila de suporte",
        "description": "MFA e RBAC granular; ações sensíveis exigem step-up e motivo.",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketPage"
                }
              }
            }
          },
          "default": {
            "description": "Erro sanitizado; códigos e condições descritos em CONVENCOES.md.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-cortes_session"
      },
      "serviceAuth": {
        "type": "mutualTLS",
        "description": "Workload autenticado no ingresso privado; sem acesso público."
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "code": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId"
        ]
      },
...
      "Ack": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "accepted": {
            "type": "boolean"
          }
        },
        "required": [
          "accepted"
        ]
      },
...
      "Account": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneMasked": {
            "type": "string"
          },
          "cpfMasked": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "BLOCKED",
              "DELETION_PENDING",
              "DELETED"
            ]
          },
          "emailVerified": {
            "type": "boolean"
          },
          "phoneVerified": {
            "type": "boolean"
          },
          "version": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "phoneMasked",
          "cpfMasked",
          "status",
          "emailVerified",
          "phoneVerified",
          "version"
        ]
      },
...
      "Package": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string"
          },
          "purchasedCredits": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "bonusCredits": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "totalCredits": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "priceMinor": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "currency": {
            "type": "string",
            "enum": [
              "BRL"
            ]
          }
        },
        "required": [
          "id",
          "code",
          "purchasedCredits",
          "bonusCredits",
          "totalCredits",
          "priceMinor",
          "currency"
        ]
      },
...
      "PurchaseInput": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "packageId": {
            "type": "string",
            "format": "uuid"
          },
          "method": {
            "type": "string",
            "enum": [
              "PIX",
              "CARD"
            ]
          },
          "paymentToken": {
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "packageId",
          "method"
        ]
      },
      "Purchase": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED",
              "CANCELLED",
              "EXPIRED",
              "PARTIALLY_REFUNDED",
              "REFUNDED",
              "CHARGEBACK"
            ]
          },
          "amountMinor": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "currency": {
            "type": "string",
            "enum": [
              "BRL"
            ]
          },
          "package": {
            "$ref": "#/components/schemas/Package"
          },
          "checkoutUrl": {
            "type": "string",
            "format": "uri"
          },
          "pixCopyPaste": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "status",
          "amountMinor",
          "currency",
          "package"
        ]
      },
...
      "TicketInput": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "subject": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10000
          },
          "category": {
            "type": "string",
            "enum": [
              "PAYMENT",
              "CREDITS",
              "ACCOUNT",
              "VIDEO",
              "CLIP",
              "SUBTITLE",
              "TECHNICAL_ERROR",
              "SUGGESTION"
            ]
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10000
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "clipId": {
            "type": "string",
            "format": "uuid"
          },
          "attachmentIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "subject",
          "category",
          "message"
        ]
      },
      "Ticket": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "subject": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "IN_REVIEW",
              "ANSWERED",
              "RESOLVED"
            ]
          },
          "projectId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "clipId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "subject",
          "category",
          "status",
          "projectId",
          "clipId",
          "createdAt"
        ]
      },
...
      "BlockInput": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "blocked": {
            "type": "boolean"
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10000
          }
        },
        "required": [
          "blocked",
          "reason"
        ]
      },
      "AdminJob": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "runId": {
            "type": "string",
            "format": "uuid"
          },
          "stage": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "attempts": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "deadlineAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastErrorCode": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "runId",
          "stage",
          "status",
          "attempts",
          "deadlineAt"
        ]
      },
...
      "PurchasePage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Purchase"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
      "TicketPage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Ticket"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
...
      "AdminJobPage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminJob"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
      "AccountPage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Account"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
...
    }
  }
}